moving Clifford class from synthesis utils to quantum_info/clifford.rs
also moving Clifford::from_gate_sequence to a standalone function clifford_from_gate_sequence
