+++ crates/quantum_info/src/lib.rs
@@ -10,6 +10,7 @@
 // copyright notice, and modified files need to carry a notice indicating
 // that they have been altered from the originals.
 
+pub mod clifford;
 pub mod convert_2q_block_matrix;
 pub mod pauli_lindblad_map;
 pub mod sparse_observable;
